+++ NEWS.md
@@ -44,7 +44,7 @@
   model classes can be defined by subclassing `stk_model_gn`.
   ([ticket #38](https://sourceforge.net/p/kriging/tickets/38/))
   
-* `stk_example_misc06.m`: Demonstrate the use of noise model objects
+* `stk_example_misc05.m`: Demonstrate the use of noise model objects
   to estimate the dispersion parameter in an heteroscedatic case.
 
 ## `stk_dataframe` and related classes